Remove jambase from frontend, refactor jambase on backend
