+++ backend/src/helpers/createPlaylist.ts
@@ -8,93 +8,6 @@ import { HttpRequestError } from "../http/HttpRequestError";
 import { Cities } from "../enums/Cities";
 import { Festivals } from "../enums/Festivals";
 
-export const CreateNewPlaylistJamBase = async ({
-  token,
-  city,
-  user_id,
-  numTopTracks,
-  spotifyIds,
-}: {
-  token: string;
-  city: string;
-  user_id: string;
-  numTopTracks?: number;
-  spotifyIds: string[];
-}) => {
-  const playlist_data: SpotifyPlaylistDataType = await CreateBlankPlaylist({
-    token,
-    city,
-    user_id,
-  });
-
-  const playlist_id = playlist_data.new_playlist_id || "";
-  try {
-    AddCoverArt({ token, playlist_id });
-  } catch (err) {
-    console.log("Error adding cover art");
-  }
-
-  const numTopTracksToAdd = numTopTracks ? numTopTracks : 1;
-
-  const promises = spotifyIds.map(async (spotifyId) => {
-    const topTracks: any = await getTopTracks(spotifyId, token);
-    let tracks = "";
-    try {
-      for (let i = 0; i < numTopTracksToAdd; i++) {
-        if (topTracks && topTracks[i]) {
-          tracks += topTracks[i];
-          tracks += ",";
-        }
-      }
-    } catch (error) {
-      //console.log(error);
-    }
-    return tracks;
-  });
-  const trackResults = await Promise.all(promises);
-
-  let tracks = trackResults.join("");
-
-  tracks = tracks.substring(0, tracks.length - 1); // remove last comma
-
-  const array = tracks.split(",");
-  const MAX_CHUNK_LENGTH = 75;
-  const trackArrays = [];
-
-  while (array.length > 0) {
-    const chunk = array.splice(0, MAX_CHUNK_LENGTH);
-    trackArrays.push(chunk);
-  }
-
-  for (const trackArray of trackArrays) {
-    const tracks = trackArray.join(",");
-    if (playlist_data.new_playlist_id && playlist_data.external_urls?.spotify) {
-      await AddTracksToPlaylist(token, playlist_data.new_playlist_id, tracks);
-    }
-  }
-  return playlist_data.external_urls?.spotify;
-};
-
-async function getTopTracks(spotifyId: string, token: string) {
-  try {
-    const response = await get({
-      url: `https://api.spotify.com/v1/artists/${spotifyId}/top-tracks?market=CA`,
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    if (response.status === 200) {
-      return response.data.tracks.map((track: any) => track.uri);
-    } else {
-      //console.log('Unexpected response status:', response.status);
-    }
-  } catch (error) {
-    //console.log('Error at fetching top tracks:', error);
-  }
-  return null; // Return null in case of errors
-}
-
 export const CreateNewPlaylist = async ({
   token,
   city,
@@ -169,7 +82,7 @@ export const CreateNewPlaylist = async ({
   return playlist_data.external_urls?.spotify;
 };
 
-const CreateBlankPlaylist = async ({
+export const CreateBlankPlaylist = async ({
   token,
   city,
   user_id,
@@ -238,7 +151,7 @@ const CreateBlankPlaylist = async ({
     });
 };
 
-const AddCoverArt = async ({ token, playlist_id }: { token: string; playlist_id: string }) => {
+export const AddCoverArt = async ({ token, playlist_id }: { token: string; playlist_id: string }) => {
   return put({
     url: "https://api.spotify.com/v1/playlists/" + playlist_id + "/images",
     headers: {
@@ -255,7 +168,7 @@ const AddCoverArt = async ({ token, playlist_id }: { token: string; playlist_id:
   });
 };
 
-const AddTracksToPlaylist = async (token: string, playlist_id: string, tracks: string) => {
+export const AddTracksToPlaylist = async (token: string, playlist_id: string, tracks: string) => {
   return post({
     url: "https://api.spotify.com/v1/playlists/" + playlist_id + "/tracks?uris=" + tracks,
     headers: {
