+++ backend/src/types/jamBase/JamBase.ts
