+++ client/src/routes/AppRoutes.tsx
@@ -2,15 +2,11 @@ import { Routes, Route } from "react-router-dom";
 import LoginPage from "../pages/LoginPage";
 import Navbarr from "../components/Navbar";
 import NotFound from "../pages/NotFound";
-import Refresh from "../pages/Refresh";
 import { Box } from "@mui/material";
-import { About } from "../pages/About";
 import { ArtistsPage } from "../pages/ArtistsPage";
 import useSpotifyAuth from "../hooks/useSpotifyAuth";
-import { LandingPage } from "../pages/LandingPage";
 import useAnalytics from "../hooks/useAnalytics";
 import { Rifflandia } from "../Rifflandia/pages/Rifflandia";
-import { JamBase } from "../pages/JamBase";
 import { SampleFestival } from "../festivals/SampleFestival/SampleFestival";
 
 export const AppRoutes = () => {
@@ -25,8 +21,6 @@ export const AppRoutes = () => {
             <Route index element={!token || !spotifyInfo || !spotifyInfo.access ? <LoginPage /> : <ArtistsPage />} />
             <Route path="/artists" element={<ArtistsPage />} />
             {/* <Route path="/about" element={<About />} /> */}
-            {/* <Route path="/refresh" element={<Refresh />} /> */}
-            <Route path="/explore" element={<JamBase />} />
             <Route path="*" element={<NotFound />} />
           </Route>
           <Route path="/rifflandia" element={<Rifflandia />} />
